% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanAnalysis.R
\docType{class}
\name{BayeScanAnalysis-class}
\alias{BayeScanAnalysis-class}
\title{BayeScanAnalysis: An S4 class to represent inputs and outputs from BayeScan}
\description{
This class stores input data and associated output results from the BayeScan program.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{BayeScanOpts} object with parameters used to run BayeScan .}

\item{\code{data}}{\code{BayeScanData} object with input data used for analysis.}

\item{\code{results}}{\code{BayeScanResults} object with results from analysis.}
}}
\seealso{
\code{\link{BayeScanAnalysis}}.
}

