% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanAnalysis.R
\name{BayeScanAnalysis}
\alias{BayeScanAnalysis}
\title{Create BayeScanAnalysis object}
\usage{
BayeScanAnalysis(opts, data, results)
}
\arguments{
\item{opts}{\code{BayeScanOpts} object with parameters used to run BayeScan .}

\item{data}{\code{BayeScanData} object with input data used for analysis.}

\item{results}{\code{BayeScanResults} object with results from analysis.}
}
\description{
This function creates a new \code{BayeScanAnalysis} object.
}
\seealso{
\code{\link{BayeScanAnalysis-class}}, \code{\link{BayeScanData}}, \code{\link{BayeScanData}}, \code{\link{BayeScanResults}}.
}

