% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\docType{class}
\name{BayeScanResults-class}
\alias{BayeScanResults-class}
\title{BayeScanResults: An S4 class to results from BayeScan}
\description{
This class stores results from the BayeScan program.
}
\section{Slots}{

\describe{
\item{\code{fst}}{\code{data.frame} object containing output results ('filename_fst.txt').}

\item{\code{mcmc}}{\code{data.frame} object containing Markov chain Monte Carlo information ('filename.sel').}

\item{\code{acceptance.rate}}{\code{data.frame} object with information on the evolution of the acceptance rate ('filename_AccRte.txt').}

\item{\code{verification}}{\code{character} object with verification diagnostics ('file_Verif.txt').}
}}
\seealso{
\code{\link{BayeScanResults}}.
}

