% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanResults.R, R/BayeScanAnalysis.R
\name{gelman.diag}
\alias{gelman.diag}
\alias{gelman.diag.BayeScanAnalysis}
\alias{gelman.diag.BayeScanResults}
\alias{gelman.diag.default}
\title{Gelman-Rubin diagnostic statistics}
\usage{
gelman.diag(x, ...)

\method{gelman.diag}{default}(x, ...)

\method{gelman.diag}{BayeScanResults}(x, ...)

\method{gelman.diag}{BayeScanAnalysis}(x, ...)
}
\arguments{
\item{x}{\code{BayeScanAnalysis} or \code{BayeScanResults} object.}

\item{...}{arguments passed to \code{\link[code]{gelman.dig}}.}
}
\description{
This function returns the Gelman-Rubin diagnostic statistics for the negative loglikelihood of multiple BayeScan runs. See \code{\link[code]{gelman.dig}} for more information.
}

