% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanReplicate.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{n.loci}
\alias{n.loci}
\alias{n.loci.BayeScanAnalysis}
\alias{n.loci.BayeScanData}
\alias{n.loci.BayeScanReplicate}
\alias{n.loci.BayeScanResults}
\title{Number of loci}
\usage{
n.loci(x)

\method{n.loci}{BayeScanReplicate}(x)

\method{n.loci}{BayeScanResults}(x)

\method{n.loci}{BayeScanData}(x)

\method{n.loci}{BayeScanAnalysis}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{integer}.
}
\description{
This function returns the number of loci in a \code{BayeScan} object.
}

