% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{n.pop}
\alias{n.pop}
\alias{n.pop.BayeScanAnalysis}
\alias{n.pop.BayeScanData}
\alias{n.pop.BayeScanResults}
\title{Number of populations}
\usage{
n.pop(x)

\method{n.pop}{BayeScanResults}(x)

\method{n.pop}{BayeScanData}(x)

\method{n.pop}{BayeScanAnalysis}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{integer}.
}
\description{
This function returns the number of populations in a \code{BayeScan} object.
}

