% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanReplicate.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{n.samples}
\alias{n.samples}
\alias{n.samples.BayeScanAnalysis}
\alias{n.samples.BayeScanData}
\alias{n.samples.BayeScanReplicate}
\alias{n.samples.BayeScanResults}
\title{Number of samples}
\usage{
n.samples(x)

\method{n.samples}{BayeScanReplicate}(x)

\method{n.samples}{BayeScanResults}(x)

\method{n.samples}{BayeScanData}(x)

\method{n.samples}{BayeScanAnalysis}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{integer}.
}
\description{
This function returns the number of samples in a \code{BayeScan} object.
}

