% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{nmds}
\alias{nmds}
\alias{nmds.BayeScanAnalysis}
\alias{nmds.BayeScanData}
\title{Nonmetric multidimensional scaling for AFLP data}
\usage{
nmds(x, max.stress, min.k, max.k, metric, type, ...)

\method{nmds}{BayeScanData}(x, max.stress = 0.1, min.k = 2, max.k = Inf,
  metric = "gower", type = "all", ...)

\method{nmds}{BayeScanAnalysis}(x, max.stress = 0.1, min.k = 2,
  max.k = Inf, metric = "gower", type = "all", ...)
}
\arguments{
\item{x}{\code{BayeScanData} object.}

\item{max.stress}{\code{numeric} maximum allowed stress. Defaults to 0.1.}

\item{min.k}{\code{numeric} minimum number of dimensions to investigate. Defaults to 2.}

\item{max.k}{\code{numeric} maximum number of dimensions to investigate. Defaults to \code{Infinity}.}

\item{metric}{\code{character} name of distance metric to use. Valid arguments are 'euclidean', 'manhattan', or 'gower'. Defaults to 'gower'.}

\item{type}{\code{character} type of loci to analyse. Valid arguments are 'all', 'adaptive', or 'neutral'. Defaults to 'all'.}

\item{...}{arguments passed to \code{link[vegan]{metaMDS}}.}
}
\value{
\code{\link[vegan]{metaMDS}} object.
}
\description{
This function performs successive nonmetric multidimensional scaling analyses on the loci in a \code{BayeScan} object until an adequate stress level is acheived.
}
\details{
The \code{link[cluster]{daisy}} function is used to calculate distances because it can accomodate missing values.
}
\examples{
# run BayeScan using low number of iterations
dat <- read.BayeScanData(system.file('extdata', 'example_fstat_aflp.dat', package='bayescanr'))
x <- run.BayeScan(dat, threads=1, n=50, thin=1, nbp=10, pilot=10, burn=10)
# generate nmds using minimal iterations and only investigating a few k values
z <- nmds(x, metric='gower', min.k=2, max.k=5, trymax=5)
}
\seealso{
\code{link[cluster]{daisy}}, \code{link[vegan]{metaMDS}}.
}

