% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R, R/generics.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{pop.names.BayeScanResults}
\alias{pop.names}
\alias{pop.names.BayeScanAnalysis}
\alias{pop.names.BayeScanData}
\alias{pop.names.BayeScanResults}
\title{Names of populations}
\usage{
\method{pop.names}{BayeScanResults}(x)

pop.names(x)

\method{pop.names}{BayeScanData}(x)

\method{pop.names}{BayeScanAnalysis}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{character}
}
\description{
This function returns the unique population names in a \code{BayeScan} object.
}

