% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanReplicate.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanOpts.R, R/BayeScanAnalysis.R
\name{print}
\alias{print}
\alias{print.BayeScanAnalysis}
\alias{print.BayeScanData}
\alias{print.BayeScanOpts}
\alias{print.BayeScanReplicate}
\alias{print.BayeScanResults}
\title{Print objects}
\usage{
\method{print}{BayeScanReplicate}(x, ..., header = TRUE)

\method{print}{BayeScanResults}(x, ..., header = TRUE)

\method{print}{BayeScanData}(x, ..., header = TRUE)

\method{print}{BayeScanOpts}(x, ..., header = TRUE)

\method{print}{BayeScanAnalysis}(x, ..., header = TRUE)
}
\arguments{
\item{x}{the object to print.}

\item{...}{not used.}

\item{header}{\code{logical} should object header be shown?}
}
\description{
This function prints objects.
}

