% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanReplicate.R
\name{read.BayeScanReplicate}
\alias{read.BayeScanReplicate}
\title{Read BayeScan run}
\usage{
read.BayeScanReplicate(file, dir, fdr = 0.95)
}
\arguments{
\item{file}{\code{character} file path of input file.}

\item{dir}{\code{dir} directory with output files.}

\item{fdr}{\code{numeric} false discovery rate threshold to classify loci as adaptive.}
}
\value{
\code{\link{BayeScanReplicate}}.
}
\description{
This function reads the results of a single run of the BayeScan program.
}
\seealso{
\code{\link{BayeScanReplicate-class}}.
}

