% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R
\name{sample.labels}
\alias{sample.labels}
\alias{sample.labels.BayeScanData}
\alias{sample.labels<-.BayeScanData}
\title{Sample labels}
\usage{
sample.labels(x)

\method{sample.labels}{BayeScanData}(x)

\method{sample.labels}{BayeScanData}(x) <- value
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}

\item{value}{\code{character} new labels.}
}
\value{
\code{character}.
}
\description{
This function returns the labels for each sample in a \code{BayeScan} object.
}

