% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanReplicate.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanAnalysis.R
\name{sample.pops}
\alias{sample.labels<-}
\alias{sample.pops}
\alias{sample.pops.BayeScanAnalysis}
\alias{sample.pops.BayeScanData}
\alias{sample.pops.BayeScanReplicate}
\alias{sample.pops.BayeScanResults}
\alias{sample.pops<-}
\alias{sample.pops<-.BayeScanData}
\title{Name of sample populations}
\usage{
sample.pops(x)

sample.pops(x) <- value

sample.labels(x) <- value

\method{sample.pops}{BayeScanReplicate}(x)

\method{sample.pops}{BayeScanResults}(x)

\method{sample.pops}{BayeScanData}(x)

\method{sample.pops}{BayeScanData}(x) <- value

\method{sample.pops}{BayeScanAnalysis}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}

\item{value}{\code{character} new population names.}
}
\value{
\code{character}.
}
\description{
This function returns the population name that each sample belongs to in a \code{BayeScan} object.
}

