% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R
\name{sample.subset}
\alias{sample.subset}
\alias{sample.subset.BayeScanData}
\title{Subset samples}
\usage{
sample.subset(x, samples)

\method{sample.subset}{BayeScanData}(x, samples)
}
\arguments{
\item{x}{\code{BayeScanData}, or \code{BayeScan}.}

\item{samples}{\code{numeric} index of samples to subset.}
}
\value{
\code{BayeScanData}.
}
\description{
This function returns a subset of the samples in a \code{BayeScan} object.
}

