% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanReplicate.R, R/BayeScanResults.R, R/BayeScanData.R, R/BayeScanOpts.R, R/BayeScanAnalysis.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BayeScanAnalysis-method}
\alias{show,BayeScanData-method}
\alias{show,BayeScanOpts-method}
\alias{show,BayeScanReplicate-method}
\alias{show,BayeScanResults-method}
\title{Show objects}
\usage{
\S4method{show}{BayeScanReplicate}(object)

\S4method{show}{BayeScanResults}(object)

\S4method{show}{BayeScanData}(object)

\S4method{show}{BayeScanOpts}(object)

\S4method{show}{BayeScanAnalysis}(object)
}
\arguments{
\item{object}{the object to show.}
}
\description{
This function shows objects.
}

