% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanResults.R, R/BayeScanAnalysis.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.BayeScanAnalysis}
\alias{traceplot.BayeScanResults}
\title{Traceplot}
\usage{
traceplot(x, ...)

\method{traceplot}{BayeScanResults}(x, ...)

\method{traceplot}{BayeScanResults}(x, ...)
}
\arguments{
\item{x}{\code{BayeScanAnalysis} or \code{BayeScanResults} object.}

\item{...}{not used.}
}
\description{
This function makes a traceplot showing convergence for multiple BayeScan runs.
}

