% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\docType{class}
\name{BayeScanResults-class}
\alias{BayeScanResults-class}
\title{BayeScanResults: An S4 class to results from BayeScan}
\description{
This class stores results from the BayeScan program.
}
\section{Slots}{

\describe{
\item{\code{summary}}{\code{data.frame} object containing overall results from BayeScan replicates.}

\item{\code{replicates}}{\code{list} of \code{BayeScanReplicate} objects.}
}}
\seealso{
\code{\link{BayeScanResults}}, \code{\link{BayeScanReplicate}}.
}

