% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\docType{class}
\name{BayeScanResults-class}
\alias{BayeScanResults-class}
\title{BayeScanResults: An S4 class to results from BayeScan}
\description{
This class stores results from the BayeScan program.
}
\section{Slots}{

\describe{
\item{\code{fst}}{\code{data.frame}}

\item{\code{mcmc}}{\code{data.frame}}

\item{\code{acceptance.rate}}{\code{data.frame}}

\item{\code{verification}}{\code{data.frame}}
}}
\seealso{
\code{\link{BayeScanResults}}.
}

