% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\name{BayeScanResults}
\alias{BayeScanResults}
\title{Create BayeScanResults object}
\usage{
BayeScanResults(summary = NULL, replicates)
}
\arguments{
\item{summary}{\code{data.frame} object containing overall results from BayeScan replicates.}

\item{replicates}{\code{list} of \code{BayeScanReplicate} objects.}
}
\value{
\code{\link{BayeScanReplicate}}.
}
\description{
This function creates a new \code{BayeScanResults} object.
}
\seealso{
\code{\link{BayeScanReplicate-class}}.
}

